courses/requirements.txt
google-genai
soundfile
google-cloud-storage
google-cloud-pubsub
functions-framework